Refactor : 가게 Repository 구조 변경
