+++ src/main/java/com/driven/dm/cart/application/service/CartService.java
@@ -18,7 +18,7 @@ import com.driven.dm.menu.domain.entity.Menu;
 import com.driven.dm.menu.infrastructure.repository.MenuRepository;
 import com.driven.dm.shop.application.exception.ShopErrorCode;
 import com.driven.dm.shop.domain.entity.Shop;
-import com.driven.dm.shop.domain.repository.ShopRepository;
+import com.driven.dm.shop.infrastructure.repository.ShopRepository;
 import com.driven.dm.user.application.exception.UserErrorCode;
 import com.driven.dm.user.domain.entity.User;
 import com.driven.dm.user.infrastructure.repository.UserRepository;
@@ -53,7 +53,7 @@ public class CartService {
         User user = userRepository.findById(userId)
             .orElseThrow(() -> new AppException(UserErrorCode.USER_NOT_FOUND));
 
-        Shop shop = shopRepository.selectShop(shopId)
+        Shop shop = shopRepository.findById(shopId)
             .orElseThrow(() -> new AppException(ShopErrorCode.SHOP_NOT_FOUND));
         if (shop == null || shop.getDeletedAt() != null) {
             throw new AppException(ShopErrorCode.SHOP_NOT_FOUND);
@@ -93,7 +93,7 @@ public class CartService {
 
         long cartTotal = cartReadRepository.sumShopCartTotal(userId, shopId);
 
-        Shop shop = shopRepository.selectShop(shopId)
+        Shop shop = shopRepository.findById(shopId)
             .orElseThrow(() -> new AppException(ShopErrorCode.SHOP_NOT_FOUND));
         if (shop == null || shop.getDeletedAt() != null) {
             throw new AppException(ShopErrorCode.SHOP_NOT_FOUND);
